:bricks: Add some helpers to JsonTextConverter
- Move JsonTextReplacement.replaceText to JsonTextConverter. The
  implementation is the same between Pf2e and Dnd, so keep this as a
  default implementation to avoid repetition, and so that callers can
  use JsonTextConverter as the base type without casting to
  Pf2eTextConverter
- Add a streamPropsExcluding helper method which provides a stream of
  properties from a JsonObject, excluding the given keys
